aot: fix trampoline re-encode condition/offset and LLVM do_exit usize cast
Two bugs introduced by the tb_ptr pointer-based refactor:

1. exec/src/exec_loop.rs - AOT trampoline had wrong branch direction and
   off-by-one skip offset:
   - Old: `cmp rax, 7; jbe +0x10` — skipped re-encoding for rax <= 7, but
     AOT functions return bare slots 0/1/2 (tb_ptr=0 at compile time) which
     need re-encoding; real exits are >= 3 (ECALL=3, EBREAK=4, UNDEF=5).
   - New: `cmp rax, 2; ja +0x11` — re-encodes slots 0/1/2 with correct
     tb_ptr, passes real exits (>2) through unchanged.
   - Skip offset fixed from 0x10 to 0x11: re-encode body is
     and(4) + movabs(10) + or(3) = 17 bytes.

2. backend/src/llvm/translate.rs - do_exit passed encode_tb_exit() result
   (usize) to ci() which expects u64; fixed with `as u64` cast.

Result: coremark-riscv64 AOT run: 5168 it/s (vs ~3100 JIT-only), nochain
exits drop from 109M to ~491 confirming all hot TBs are resolved within
the AOT .so via musttail chaining.

+++ backend/src/llvm/translate.rs
@@ -356,7 +356,7 @@ impl TbTranslator {
     fn do_exit(&self, exit_code: u64) {
         let encoded = tcg_core::tb::encode_tb_exit(self.tb_ptr, exit_code);
         unsafe {
-            LLVMBuildStore(self.builder, self.ci(self.i64t, encoded), self.exit_val);
+            LLVMBuildStore(self.builder, self.ci(self.i64t, encoded as u64), self.exit_val);
             LLVMBuildBr(self.builder, self.exit_bb);
         }
     }
